Move OOP lsp server in proc and unify search implementations
