+++ src/Features/LanguageServer/ProtocolUnitTests/Symbols/WorkspaceSymbolsTests.cs
@@ -2,12 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#nullable disable
-
 using System;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces;
 using Microsoft.CodeAnalysis.LanguageServer.Handler;
 using Roslyn.Test.Utilities;
 using Xunit;
@@ -30,7 +29,7 @@ namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.Symbols
             using var workspace = CreateTestWorkspace(markup, out var locations);
             var expected = new LSP.SymbolInformation[]
             {
-                CreateSymbolInformation(LSP.SymbolKind.Class, "A", locations["class"].Single())
+                CreateSymbolInformation(LSP.SymbolKind.Class, "A", locations["class"].Single(), Glyph.ClassInternal, GetContainerName(workspace))
             };
 
             var results = await RunGetWorkspaceSymbolsAsync(workspace.CurrentSolution, "A").ConfigureAwait(false);
@@ -50,14 +49,14 @@ namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.Symbols
             using var workspace = CreateTestWorkspace(markup, out var locations);
             var expected = new LSP.SymbolInformation[]
             {
-                CreateSymbolInformation(LSP.SymbolKind.Class, "A", locations["class"].Single())
+                CreateSymbolInformation(LSP.SymbolKind.Class, "A", locations["class"].Single(), Glyph.ClassInternal, GetContainerName(workspace))
             };
 
             using var progress = BufferedProgress.Create<LSP.SymbolInformation>(null);
 
             var results = await RunGetWorkspaceSymbolsAsync(workspace.CurrentSolution, "A", progress).ConfigureAwait(false);
 
-            Assert.Empty(results);
+            Assert.Null(results);
 
             results = progress.GetValues().ToArray();
 
@@ -77,7 +76,7 @@ namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.Symbols
             using var workspace = CreateTestWorkspace(markup, out var locations);
             var expected = new LSP.SymbolInformation[]
             {
-                CreateSymbolInformation(LSP.SymbolKind.Method, "M", locations["method"].Single())
+                CreateSymbolInformation(LSP.SymbolKind.Method, "M", locations["method"].Single(), Glyph.MethodPrivate, GetContainerName(workspace, "A"))
             };
 
             var results = await RunGetWorkspaceSymbolsAsync(workspace.CurrentSolution, "M").ConfigureAwait(false);
@@ -100,7 +99,7 @@ namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.Symbols
             using var workspace = CreateTestWorkspace(markup, out var locations);
             var expected = new LSP.SymbolInformation[]
             {
-                CreateSymbolInformation(LSP.SymbolKind.Variable, "i", locations["local"].Single())
+                CreateSymbolInformation(LSP.SymbolKind.Variable, "i", locations["local"].Single(), Glyph.Local, GetContainerName(workspace, "A.M.i"))
             };
 
             var results = await RunGetWorkspaceSymbolsAsync(workspace.CurrentSolution, "i").ConfigureAwait(false);
@@ -123,11 +122,12 @@ namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.Symbols
     }
 }";
             using var workspace = CreateTestWorkspace(markup, out var locations);
+            var classAContainerName = GetContainerName(workspace, "A");
             var expected = new LSP.SymbolInformation[]
             {
-                CreateSymbolInformation(LSP.SymbolKind.Field, "F", locations["field"][0]),
-                CreateSymbolInformation(LSP.SymbolKind.Class, "F", locations["class"].Single()),
-                CreateSymbolInformation(LSP.SymbolKind.Field, "F", locations["field"][1])
+                CreateSymbolInformation(LSP.SymbolKind.Field, "F", locations["field"][0], Glyph.FieldPrivate, classAContainerName),
+                CreateSymbolInformation(LSP.SymbolKind.Class, "F", locations["class"].Single(), Glyph.ClassPrivate, classAContainerName),
+                CreateSymbolInformation(LSP.SymbolKind.Field, "F", locations["field"][1], Glyph.FieldPrivate, GetContainerName(workspace, "A.F"))
             };
 
             var results = await RunGetWorkspaceSymbolsAsync(workspace.CurrentSolution, "F").ConfigureAwait(false);
@@ -156,8 +156,8 @@ namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.Symbols
             using var workspace = CreateTestWorkspace(markups, out var locations);
             var expected = new LSP.SymbolInformation[]
             {
-                CreateSymbolInformation(LSP.SymbolKind.Method, "M", locations["method"][0]),
-                CreateSymbolInformation(LSP.SymbolKind.Method, "M", locations["method"][1])
+                CreateSymbolInformation(LSP.SymbolKind.Method, "M", locations["method"][0], Glyph.MethodPrivate, GetContainerName(workspace, "A")),
+                CreateSymbolInformation(LSP.SymbolKind.Method, "M", locations["method"][1], Glyph.MethodPrivate, GetContainerName(workspace, "B"))
             };
 
             var results = await RunGetWorkspaceSymbolsAsync(workspace.CurrentSolution, "M").ConfigureAwait(false);
@@ -180,6 +180,24 @@ namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.Symbols
             Assert.Empty(results);
         }
 
+        [Fact]
+        public async Task TestGetWorkspaceSymbolsAsync_VisualBasic()
+        {
+            var markup = @"Class {|class:A|}
+    Sub Method()
+    End Sub
+End Class";
+
+            using var workspace = CreateVisualBasicTestWorkspace(markup, out var locations);
+            var expected = new LSP.SymbolInformation[]
+            {
+                CreateSymbolInformation(LSP.SymbolKind.Class, "A", locations["class"].Single(), Glyph.ClassInternal, GetContainerName(workspace))
+            };
+
+            var results = await RunGetWorkspaceSymbolsAsync(workspace.CurrentSolution, "A").ConfigureAwait(false);
+            AssertJsonEquals(expected, results);
+        }
+
         private static async Task<LSP.SymbolInformation[]> RunGetWorkspaceSymbolsAsync(Solution solution, string query, IProgress<LSP.SymbolInformation[]> progress = null)
         {
             var request = new LSP.WorkspaceSymbolParams
@@ -191,6 +209,18 @@ namespace Microsoft.CodeAnalysis.LanguageServer.UnitTests.Symbols
             var queue = CreateRequestQueue(solution);
             return await GetLanguageServer(solution).ExecuteRequestAsync<LSP.WorkspaceSymbolParams, LSP.SymbolInformation[]>(queue, LSP.Methods.WorkspaceSymbolName,
                 request, new LSP.ClientCapabilities(), null, CancellationToken.None);
+        }
+
+        private static string GetContainerName(TestWorkspace testWorkspace, string? containingSymbolName = null)
+        {
+            if (containingSymbolName == null)
+            {
+                return string.Format(FeaturesResources.project_0, testWorkspace.Projects.Single().Name);
+            }
+            else
+            {
+                return string.Format(FeaturesResources.in_0_project_1, containingSymbolName, testWorkspace.Projects.Single().Name);
+            }
         }
     }
 }
