+++ src/VisualStudio/Core/Def/Implementation/LanguageClient/VisualStudioConnectionEventListener.cs
@@ -23,7 +23,7 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.LanguageClient
     [ExportEventListener(WellKnownEventListeners.Workspace, WorkspaceKind.Host), Shared]
     internal class VisualStudioOOPConnectionEventListener : IEventListener<object>
     {
-        private readonly VisualStudioOOPConnectionLanguageClient _languageClient;
+        private readonly VisualStudioInProcLanguageClient _languageClient;
         private readonly Lazy<ILanguageClientBroker> _languageClientBroker;
 
         private readonly IAsynchronousOperationListener _asynchronousOperationListener;
@@ -31,7 +31,7 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.LanguageClient
         [ImportingConstructor]
         [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
         public VisualStudioOOPConnectionEventListener(
-            VisualStudioOOPConnectionLanguageClient languageClient,
+            VisualStudioInProcLanguageClient languageClient,
             Lazy<ILanguageClientBroker> languageClientBroker,
             IAsynchronousOperationListenerProvider listenerProvider)
         {
@@ -42,7 +42,7 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.LanguageClient
 
         /// <summary>
         /// LSP clients do not necessarily know which language servers (and when) to activate as they are language
-        /// agnostic.  We know we can provide <see cref="VisualStudioOOPConnectionLanguageClient"/> as soon as the
+        /// agnostic.  We know we can provide <see cref="VisualStudioInProcLanguageClient"/> as soon as the
         /// workspace is started, so tell the <see cref="ILanguageClientBroker"/> to start loading it.
         /// </summary>
         public void StartListening(Workspace workspace, object serviceOpt)
