+++ src/Workspaces/Core/Portable/Remote/WellKnownServiceHubService.cs
@@ -13,7 +13,7 @@ namespace Microsoft.CodeAnalysis.Remote
         // obsolete: RemoteDesignerAttributeService = 4,
         // obsolete: RemoteProjectTelemetryService = 5,
         // obsolete: RemoteTodoCommentsService = 6,
-        RemoteLanguageServer = 7,
+        // obsolete: RemoteLanguageServer = 7,
         IntelliCode = 8,
         Razor = 9,
